test, move reward distribution test in e2e test framework.
