--- consensus/test/ac_reward_distribution_test.go
@@ -1,10 +1,0 @@
-package test
-
-import (
-	"testing"
-)
-
-// todo: at the end of each epoch, do reward distribution checks.
-func TestRewardDistribution(t *testing.T) {
-	t.Skip("to be done in epoch approach")
-}
